pyproject.toml
[project]
name = "pelican"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "pillow>=12.1.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
]

[project.scripts]
pelican = "timmins:hello_world"